+++ service-tree/src/test/java/com/zhuinden/servicetree/TraversalTest.java
@@ -103,11 +103,38 @@ public class TraversalTest {
                 }
             }
         });
-        assertThat(nodes).containsExactly(root1Node, child1ANode, child1A1Node, child1A2Node, child1BNode, child1B1Node);
+        assertThat(nodes).containsExactly(root1Node,
+                child1ANode,
+                child1A1Node,
+                child1A2Node,
+                child1BNode,
+                child1B1Node);
     }
 
     @Test
     public void traversePostOrderWorksAsIntended() {
+        final List<ServiceTree.Node> nodes = new LinkedList<>();
+        serviceTree.traverseTree(ServiceTree.Walk.POST_ORDER, new ServiceTree.Walk() {
+            @Override
+            public void execute(@NonNull ServiceTree.Node node, @NonNull CancellationToken cancellationToken) {
+                nodes.add(node);
+            }
+        });
+        assertThat(nodes).containsExactly( //
+                child2BNode,
+                child2ANode,
+                root2Node,
+                child1B2Node,
+                child1B1Node,
+                child1BNode,
+                child1A2Node,
+                child1A1Node,
+                child1ANode,
+                root1Node);
+    }
+
+    @Test
+    public void traversePostOrderCancelWorksAsIntended() {
         final List<ServiceTree.Node> nodes = new LinkedList<>();
         serviceTree.traverseTree(ServiceTree.Walk.POST_ORDER, new ServiceTree.Walk() {
             @Override
@@ -119,7 +146,14 @@ public class TraversalTest {
             }
         });
         assertThat(nodes).containsExactly( //
-                child2BNode, child2ANode, root2Node, child1B2Node, child1B1Node, child1BNode, child1A2Node, child1A1Node);
+                child2BNode,
+                child2ANode,
+                root2Node,
+                child1B2Node,
+                child1B1Node,
+                child1BNode,
+                child1A2Node,
+                child1A1Node);
     }
 
     @Test
@@ -189,6 +223,18 @@ public class TraversalTest {
     }
 
     @Test
+    public void traverseChainPostOrderWorksAsIntended() {
+        final List<ServiceTree.Node> nodes = new LinkedList<>();
+        serviceTree.traverseChain(child1A1Node, ServiceTree.Walk.POST_ORDER, new ServiceTree.Walk() {
+            @Override
+            public void execute(@NonNull ServiceTree.Node node, @NonNull CancellationToken cancellationToken) {
+                nodes.add(node);
+            }
+        });
+        assertThat(nodes).containsExactly(child1A1Node, child1ANode, root1Node);
+    }
+
+    @Test
     public void findRootFindsRoot() {
         assertThat(serviceTree.findRoot(root1Node)).isSameAs(root1Node);
     }
@@ -196,5 +242,75 @@ public class TraversalTest {
     @Test
     public void findRootForChildFindsRoot() {
         assertThat(serviceTree.findRoot(child1A2Node)).isSameAs(root1Node);
+    }
+
+    @Test
+    public void traverseChainIncludingRootWorksAsIntended() {
+        final List<ServiceTree.Node> nodes = new LinkedList<>();
+        serviceTree.traverseChain(child1B2Node, true, new ServiceTree.Walk() {
+            @Override
+            public void execute(@NonNull ServiceTree.Node node, @NonNull CancellationToken cancellationToken) {
+                nodes.add(node);
+            }
+        });
+        assertThat(nodes).containsExactly(child1B2Node, child1BNode, root1Node, serviceTree.getTreeRoot());
+    }
+
+    @Test
+    public void traversePreOrderIncludingRootWorksAsIntended() {
+        final List<ServiceTree.Node> nodes = new LinkedList<>();
+        serviceTree.traverseTree(ServiceTree.Walk.PRE_ORDER, true, new ServiceTree.Walk() {
+            @Override
+            public void execute(@NonNull ServiceTree.Node node, @NonNull CancellationToken cancellationToken) {
+                nodes.add(node);
+            }
+        });
+        assertThat(nodes).containsExactly( //
+                serviceTree.getTreeRoot(),
+                root1Node,
+                child1ANode,
+                child1A1Node,
+                child1A2Node,
+                child1BNode,
+                child1B1Node,
+                child1B2Node,
+                root2Node,
+                child2ANode,
+                child2BNode);
+    }
+
+    @Test
+    public void traversePostOrderIncludingRootWorksAsIntended() {
+        final List<ServiceTree.Node> nodes = new LinkedList<>();
+        serviceTree.traverseTree(ServiceTree.Walk.POST_ORDER, true, new ServiceTree.Walk() {
+            @Override
+            public void execute(@NonNull ServiceTree.Node node, @NonNull CancellationToken cancellationToken) {
+                nodes.add(node);
+            }
+        });
+        assertThat(nodes).containsExactly( //
+                child2BNode,
+                child2ANode,
+                root2Node,
+                child1B2Node,
+                child1B1Node,
+                child1BNode,
+                child1A2Node,
+                child1A1Node,
+                child1ANode,
+                root1Node,
+                serviceTree.getTreeRoot());
+    }
+
+    @Test
+    public void traverseChainPostOrderIncludingRootWorksAsIntended() {
+        final List<ServiceTree.Node> nodes = new LinkedList<>();
+        serviceTree.traverseChain(child1A1Node, ServiceTree.Walk.POST_ORDER, true, new ServiceTree.Walk() {
+            @Override
+            public void execute(@NonNull ServiceTree.Node node, @NonNull CancellationToken cancellationToken) {
+                nodes.add(node);
+            }
+        });
+        assertThat(nodes).containsExactly(child1A1Node, child1ANode, root1Node, serviceTree.getTreeRoot());
     }
 }